% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_github_template_url}
\alias{set_github_template_url}
\title{Set Github Template URL}
\usage{
set_github_template_url(pkgbuild_url = NULL)
}
\arguments{
\item{pkgbuild_url}{A string identifying a GitHub file URL for "Build Package" template}
}
\value{
a list of all \code{pkgtemp} options sorted by name.
}
\description{
You can use your own template from GitHub.
Provide URL to your template file in GitHub to argument in this function.
}
